实现了功能的自定义(除了delete_some)
